Rewrite image to disk if updating timestamp failed
In some cases Android will refuse to update any timestamps unless you
are running the program with root priviliges, which obviously isn't the
case for most apps. Therefore, if the call to File.setLastModified()
fails, we manually re-save the image to disk. This is a bit slow, but
much faster than not being able to re-check the timestamp later.