Cancel late requests processed by other workers
It can happen that both workers end up processing a request for a single
listener, and one of them may deliver the wrong image to the listener.
Though this sounds quite rare, it's easy to replicate -- just flip
around back and forth really fast in the demo app. This patch makes each
worker invalidate other worker's listener in case that worker started
processing an earlier request to the same object.