Images returned as Bitmaps, not Drawables
Doing this offers a few memory and performance related advantages.
First, it is not necessary to create so many BitmapDrawable instances
wrapping the Bitmaps, especially when the common use-case is simply to
set the bitmap to an ImageView.

Furthermore, using Bitmaps allows the caller to pass in a BitmapFactory
Options instance, which can be used to do scaling or downsampling of the
image. This can improve memory usage significantly.

When images are finished downloading, they are still saved in hi-res
format and ignore any options. Also, the library has been refactored to
save the downloaded images to disk directly, and then loading them from
file directly afterwards. This allows us the opportunity to first
recycle the downloaded bitmap and resample if desired.