Save images to temporary location, move when done
Again, if multiple items in a ListView request the same item, subsequent
requests will be rerouted to the FileLoaderThread's queue. As that task
only checks for the presence of a file, it will try to load the
incomplete images from file and (most likely) return null. So it is very
important that we write the images to some other location on disk and
then move them to the expected location when the download completes.