Add forceDownload flag to ImageRequest
If we could not delete a file which had expired, then an infinite loop
would be created here. This is also much clearer as to the true
intention here -- to replace the cached file with a newer file, in which
case we should not care about the existence of the cached file.