Should set pending image on visibility change
Oye, what a dumb bug here. Basically, if the activity window is hidden
and then re-shown, we need a way to trigger reloading of any images, as
the activity's onPause() method should cancel all pending requests.
However, when that activity is shown again, any WebImageView without a
loaded image needs to re-issue that request. However, that request
should be for the previously *pending* image, not the loaded image.